@@ -18,7 +18,7 @@ class _ChartApp extends StatelessWidget {
 
 class _MyHomePage extends StatefulWidget {
   // ignore: prefer_const_constructors_in_immutables
-  _MyHomePage({Key key}) : super(key: key);
+  _MyHomePage({Key? key}) : super(key: key);
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
@@ -54,11 +54,11 @@ class _MyHomePageState extends State<_MyHomePage> {
                       ],
                       gradient: LinearGradient(
                           colors: [
-                            Colors.purpleAccent[700],
-                            Colors.purpleAccent[100],
-                            Colors.purple[600],
-                            Colors.deepPurpleAccent[400],
-                            Colors.purple[900],
+                            Colors.purpleAccent[700]!,
+                            Colors.purpleAccent[100]!,
+                            Colors.purple[600]!,
+                            Colors.deepPurpleAccent[400]!,
+                            Colors.purple[900]!,
                           ],
                           stops: <double>[
                             0.1,
